Switch to GCC 4.2 and quash some warnings (maybe fixing an MSN bug in the process?)

+++ Plugins/Purple Service/ESZephyrService.m
@@ -60,7 +60,7 @@
 - (NSCharacterSet *)ignoredCharacters{
 	return [NSCharacterSet characterSetWithCharactersInString:@""];
 }
-- (int)allowedLength{
+- (NSUInteger)allowedLength{
 	return 255;
 }
 - (BOOL)caseSensitive{
